测试

+++ pref/src/linux.rs
@@ -166,8 +166,8 @@ fn get_cpu_args(process: &process::Process) -> (u64, u64, i64, i64) {
     if let Ok(sys) = system::cpu_times() {
         return (sys.system,                                 //系统内核态cpu占用时间，单位秒
                 sys.user + sys.nice,                        //系统用户态cpu占用时间，单位秒
-                process.stime_ticks + process.cstime_ticks, //进程内核态cpu占用时间，单位秒
-                process.utime_ticks + process.cutime_ticks) //进程用户态cpu占用时间，单位秒
+                process.stime_ticks as i64 + process.cstime_ticks, //进程内核态cpu占用时间，单位秒
+                process.utime_ticks as i64 + process.cutime_ticks) //进程用户态cpu占用时间，单位秒
     }
 
     (0, 0, 0, 0)
